🩹 Fix(review): PR #35 리뷰 반영
- SyncService: renamed 파일 처리 시 deactive(file.filename) → deactive(file.previous_filename!) 버그 수정
- path-utils: computeFolderPaths() 유틸 함수 추출하여 3곳 중복 제거
  (sitemap.ts, FolderSidebarWrapper.tsx, category/[...path]/page.tsx)
