+++ src/app/components/FolderSidebarWrapper.tsx
@@ -1,6 +1,7 @@
 import { getRepositories } from "@/infra/db/repositories";
 import { categoryIcons } from "@/infra/db/constants";
 import { FolderSidebar } from "@/components/FolderSidebar";
+import { computeFolderPaths } from "@/lib/path-utils";
 
 export async function FolderSidebarWrapper() {
   try {
@@ -9,20 +10,10 @@ export async function FolderSidebarWrapper() {
       post.getAllPostPaths(),
       post.getAllPostsForSidebar(),
     ]);
-    const folderPathSet = new Set<string>();
-    for (const p of postPaths) {
-      const parts = p.split("/");
-      for (let i = 1; i < parts.length; i++) {
-        folderPathSet.add(parts.slice(0, i).join("/"));
-      }
-    }
-    const folderPaths = Array.from(folderPathSet)
-      .sort()
-      .map((p) => p.split("/"));
 
     return (
       <FolderSidebar
-        folderPaths={folderPaths}
+        folderPaths={computeFolderPaths(postPaths)}
         posts={posts}
         categoryIcons={categoryIcons}
       />
